+++ Halforbit.DataStores/Observation/IObserver.cs
@@ -1,0 +1,18 @@
+﻿using System.Threading.Tasks;
+
+namespace Halforbit.DataStores
+{
+    public interface IObserver
+    {
+        Task BeforePut(object key, object value);
+
+        Task BeforeDelete(object key);
+    }
+
+    public interface IObserver<TKey, TValue>
+    {
+        Task BeforePut(TKey key, TValue value);
+
+        Task BeforeDelete(TKey key);
+    }
+}